Add modular javascript to the repo
